refac: 구입금액 타입 Int -> Long 수정

+++ src/main/kotlin/lotto/domain/LottoMatcher.kt
@@ -15,8 +15,8 @@ class LottoMatcher() {
         return LottoMatchResult(matchedResult, earnedMoney)
     }
 
-    fun calculateEarnedRate(earnedMoney: EarnedMoney, paidMoney: Int): EarnedRate {
-        val earnedRate = earnedMoney.money.toFloat() / paidMoney
+    fun calculateEarnedRate(earnedMoney: EarnedMoney, paidMoney: Long): EarnedRate {
+        val earnedRate = earnedMoney.money / paidMoney.toFloat()
         return EarnedRate(earnedRate)
     }
