+++ src/test/kotlin/lotto/domain/LottoTicketShopTest.kt
@@ -19,13 +19,13 @@ class LottoTicketShopTest {
 
     @ParameterizedTest
     @CsvSource(value = ["14000|14", "1000|1", "2000|2"], delimiter = '|')
-    fun `구입 금액만큼 로또를 구입할 수 있다`(money: Int, expected: Int) {
+    fun `구입 금액만큼 로또를 구입할 수 있다`(money: Long, expected: Int) {
         assertThat(lottoShop.buyLotto(money).lottoCount).isEqualTo(expected)
     }
 
     @ParameterizedTest
     @ValueSource(strings = ["900", "0"])
-    fun `로또 구입 금액이 1천원 미만이면 IllegalArgumentException 예외가 발생한다`(money: Int) {
+    fun `로또 구입 금액이 1천원 미만이면 IllegalArgumentException 예외가 발생한다`(money: Long) {
         assertThrows<IllegalArgumentException> {
             lottoShop.buyLotto(money)
         }
@@ -33,7 +33,7 @@ class LottoTicketShopTest {
 
     @ParameterizedTest
     @ValueSource(strings = ["1900", "2001"])
-    fun `로또 구입 금액이 1천원 이상인데, 천단위가 아니면 IllegalArgumentException 예외가 발생한다`(money: Int) {
+    fun `로또 구입 금액이 1천원 이상인데, 천단위가 아니면 IllegalArgumentException 예외가 발생한다`(money: Long) {
         assertThrows<IllegalArgumentException> {
             lottoShop.buyLotto(money)
         }
